% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cN.R
\name{cnbinomMME}
\alias{cnbinomMME}
\title{MME estimates for the parameters r and beta in the continuous Negative Binomial distribution}
\usage{
cnbinomMME(data)
}
\arguments{
\item{data}{required to check for model fitting}
}
\value{
the estimates of r and beta for the continuous Negative Binomial distribution
}
\description{
MME estimates for the parameters r and beta in the continuous Negative Binomial distribution
}
\examples{
cnbinomMME(data = rcnbinom(n = 100, r = 2, beta = 3))
}
