% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cN.R
\name{dcnbinom}
\alias{dcnbinom}
\title{Density for the continuous Negative Binomial distribution with parameters r and beta}
\usage{
dcnbinom(x, r, beta)
}
\arguments{
\item{x}{vector of (non-negative) quantiles}

\item{r}{parameter from the hierarchical continuous gamma distribution}

\item{beta}{parameter from the hierarchical continuous beta distribution}
}
\value{
value of the probability density function at x
}
\description{
Density for the continuous Negative Binomial distribution with parameters r and beta
}
\examples{
dcnbinom(x = 1, r = 2, beta = 3)
}
