% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cN.R
\name{rcnbinom}
\alias{rcnbinom}
\title{Random number generation for the continuous Negative Binomial distribution with parameters r and beta}
\usage{
rcnbinom(n, r, beta)
}
\arguments{
\item{n}{number of observations}

\item{r}{parameter from the hierarchical gamma distribution}

\item{beta}{parameter from the hierarchical gamma distribution}
}
\value{
vector containing random values from the continuous Negative Binomial distribution with parameters r and beta
}
\description{
Random number generation for the continuous Negative Binomial distribution with parameters r and beta
}
\examples{
rcnbinom(n = 100, r = 3, beta = 7)
}
